#15 add test
