+++ src/test/java/com/uchicom/repty/AbstractTest.java
@@ -1,0 +1,20 @@
+package com.uchicom.repty;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.mockito.MockitoAnnotations;
+
+public abstract class AbstractTest {
+
+	AutoCloseable closeable;
+
+	@BeforeEach
+	public void openMocks() {
+		closeable = MockitoAnnotations.openMocks(this);
+	}
+
+	@AfterEach
+	public void releaseMocks() throws Exception {
+		closeable.close();
+	}
+}
